XXX tests: Add qr/\p{Name=...}/
This accomplishes the same thing as \N{...}, but only for regex
patterns, and uses loose matching.  This commit includes a comparison of
the two approaches, added to perlunicode.  But the real reason to do
this is as a way station to being able to specify wild card lookup on
the name property, coming in a later commit.